Write magic value to RAM before resetting system
This value can be used by the bootloader to determine if the reset had
been requested over the serial line with the DTR negative edge + "1EAF"
sequence.

New bootloader versions relying on this magic to choose between
perpetual bootloader mode or user application (instead of the previous
2s timeout) will require this modification in the sketches' libmaple,
otherwise the user button will have to be pressed in order to enter
bootloader mode.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>